Add & fix various stuff
